fix code in state
