+++ src/pages/CreateSaleCar/index.tsx
@@ -23,7 +23,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { schema, type CreateSaleCarForm } from "./validations";
 import { Congrats } from "@/components/core/containers/Congrats";
 import { Button } from "@/components/common/Button";
-import { congratsType } from "./constants";
+import { carModels, congratsType } from "./constants";
 import { createSaleCar } from "@/services/domain/cars";
 import type {
   BasicSaleCar,
@@ -32,61 +32,9 @@ import type {
 import { useCtaMutation } from "@/hooks/useCtaMutation";
 import { Breadcrumbs } from "@/components/common/Breadcrumb";
 
-// Car Models
-const carModels = [
-  {
-    id: "1",
-    name: "Tesla Model 3",
-    image:
-      "https://images.unsplash.com/photo-1560958089-b8a1929cea89?w=100&h=60&fit=crop",
-  },
-  {
-    id: "2",
-    name: "BMW Serie 3",
-    image:
-      "https://images.unsplash.com/photo-1555215695-3004980ad54e?w=100&h=60&fit=crop",
-  },
-  {
-    id: "3",
-    name: "Mercedes-Benz Clase C",
-    image:
-      "https://images.unsplash.com/photo-1618843479313-40f8afb4b4d8?w=100&h=60&fit=crop",
-  },
-  {
-    id: "4",
-    name: "Audi A4",
-    image:
-      "https://images.unsplash.com/photo-1606664515524-ed2f786a0bd6?w=100&h=60&fit=crop",
-  },
-  {
-    id: "5",
-    name: "Toyota Camry",
-    image:
-      "https://images.unsplash.com/photo-1621007947382-bb3c3994e3fb?w=100&h=60&fit=crop",
-  },
-  {
-    id: "6",
-    name: "Honda Accord",
-    image:
-      "https://images.unsplash.com/photo-1590362891991-f776e747a588?w=100&h=60&fit=crop",
-  },
-  {
-    id: "7",
-    name: "Ford Mustang",
-    image:
-      "https://images.unsplash.com/photo-1584345604476-8ec5f5f261ca?w=100&h=60&fit=crop",
-  },
-  {
-    id: "8",
-    name: "Chevrolet Camaro",
-    image:
-      "https://images.unsplash.com/photo-1552519507-da3b142c6e3d?w=100&h=60&fit=crop",
-  },
-];
-
 const CreateSaleCar = () => {
   const navigate = useNavigate();
-  const [type, setSuccess] = useState<"success" | "error" | null>(null);
+  const [type, setType] = useState<"success" | "error" | null>(null);
   const [isLoading, setIsLoading] = useState(false);
 
   const createSaleCarMutation = useCtaMutation<SaleCarResponse, BasicSaleCar>(
@@ -114,9 +62,9 @@ const CreateSaleCar = () => {
         price: Number(data.price),
       });
 
-      setSuccess("success");
+      setType("success");
     } catch (error) {
-      setSuccess("error");
+      setType("error");
     } finally {
       setIsLoading(false);
     }
@@ -234,13 +182,14 @@ const CreateSaleCar = () => {
                     label="Precio de Venta"
                     type="number"
                     disabled={isLoading}
-                    InputProps={{
-                      startAdornment: (
+                    slotProps={
+                        {input:{
+                            startAdornment:(
                         <InputAdornment position="start">
                           <MoneyIcon sx={{ color: "text.secondary" }} />
-                        </InputAdornment>
-                      ),
-                    }}
+                        </InputAdornment>)
+                        }}
+                    }
                     helperText={
                       errors.price?.message ||
                       "Ingresa el precio en tu moneda local"
